+++ code/src/nuvla/job/actions/deployment_start.py
@@ -244,7 +244,7 @@ class DeploymentStartJob(object):
         try:
             self.handle_deployment(deployment)
         except Exception as ex:
-            log.error('Failed to start {0}: {1}'.format(deployment_id, ex))
+            log.error('Failed to {0} {1}: {2}'.format(self.job['action'], deployment_id, ex))
             try:
                 self.job.set_status_message(repr(ex))
                 self.api_dpl.set_state_error(deployment_id)