+++ code/src/nuvla/job/actions/deployment_stop.py
@@ -77,7 +77,7 @@ class DeploymentStopJob(object):
             elif Deployment.is_application(deployment):
                 self.stop_application(deployment)
         except Exception as ex:
-            log.error('Failed to start {0}: {1}'.format(deployment_id, ex))
+            log.error('Failed to {0} {1}: {2}'.format(self.job['action'], deployment_id, ex))
             try:
                 self.job.set_status_message(repr(ex))
                 self.api_dpl.set_state_error(deployment_id)